Delete Student and Section work with making validating no enrollments for either before deleting

+++ main.py
@@ -139,14 +139,19 @@ def delete_student(session: Session):
     :return:            None
     """
     student: Student = select_student(session)
-    """This is a bit ghetto.  The relationship from Student to StudentMajor has 
-    cascade delete, so this delete will work even if a student has declared one
-    or more majors.  I could write a method on Student that would return some
-    indication of whether it has any children, and use that to let the user know
-    that they cannot delete this particular student.  But I'm too lazy at this
-    point.
-    """
-    session.delete(student)
+    recs = sess.query(Student).join(Enrollment, Student.studentId == Enrollment.studentId).join(
+        Section, Enrollment.sectionId == Section.sectionId).filter(
+        Student.studentId == student.studentId).add_columns(
+        Student.lastName, Student.firstName, Section.courseNumber, Section.sectionNumber).all()
+    student_enrollments: int = sess.query(Student).join(Enrollment, Student.studentId == Enrollment.studentId).filter(
+        Student.studentId == Enrollment.studentId).count()
+    if student_enrollments > 0:
+        print(f"Student has the following sections so they will not be deleted")
+        for stu in recs:
+            print(f"Student name: {stu.lastName}, {stu.firstName}, Course Number: {stu.courseNumber},"
+                  f" Section: {stu.sectionNumber}")
+    else:
+        session.delete(student)
 
 def list_student(session: Session):
     """
@@ -566,8 +571,20 @@ def select_section(sess: Session) -> Section:
 
 def delete_section(sess: Session):
     print("Deleting a section")
-    OldSection = select_section(sess)
-    sess.delete(OldSection)
+    section: Section = select_section(sess)
+    recs = sess.query(Section).join(Enrollment, Enrollment.sectionId == Section.sectionId).join(
+        Student, Enrollment.studentId == Student.studentId).filter(
+        Section.sectionId == section.sectionId).add_columns(
+        Student.lastName, Student.firstName, Section.courseNumber, Section.sectionNumber).all()
+    section_enrollments: int = sess.query(Section).join(Enrollment, Section.sectionId == Enrollment.sectionId).filter(
+        Section.sectionId == Enrollment.sectionId).count()
+    if section_enrollments > 0:
+        print(f"Section has the following students so it will not be deleted")
+        for stu in recs:
+            print(f"Student name: {stu.lastName}, {stu.firstName}, Course Number: {stu.courseNumber},"
+                  f" Section: {stu.sectionNumber}")
+    else:
+        sess.delete(section)
 
 def add_major(session: Session):
     """
@@ -750,11 +767,14 @@ def unenroll_student_section(sess):
     section: Section = select_section(sess)
     student_section_count: int = sess.query(Enrollment).filter(Enrollment.studentId == student.studentId,
                                                                Enrollment.sectionId == section.sectionId).count()
-    unique_student_section: bool = student_section_count == 0
-    while unique_student_section:
+    unique_student_section: bool = student_section_count == 1
+    while not unique_student_section:
         print("That student does not have that section.  Try again.")
         student = select_student(sess)
         section = select_section(sess)
+        student_section_count: int = sess.query(Enrollment).filter(Enrollment.studentId == student.studentId,
+                                                                   Enrollment.sectionId == section.sectionId).count()
+        unique_student_section = student_section_count == 1
     student.remove_enrollment(section)
 
 def unenroll_section_student(sess):
@@ -767,15 +787,18 @@ def unenroll_section_student(sess):
     student: Student = select_student(sess)
     student_section_count: int = sess.query(Enrollment).filter(Enrollment.studentId == student.studentId,
                                                                Enrollment.sectionId == section.sectionId).count()
-    unique_student_section: bool = student_section_count == 0
-    while unique_student_section:
+    unique_student_section: bool = student_section_count == 1
+    while not unique_student_section:
         print("That section does not have that student.  Try again.")
         section = select_section(sess)
         student = select_student(sess)
+        student_section_count: int = sess.query(Enrollment).filter(Enrollment.studentId == student.studentId,
+                                                                   Enrollment.sectionId == section.sectionId).count()
+        unique_student_section = student_section_count == 1
     section.remove_enrollment(student)
 
 def list_student_section(sess: Session):
-    """Prompt the user for the student, and then list the majors that the student has declared.
+    """Prompt the user for the student, and then list the section that the student has declared.
     :param sess:    The connection to the database
     :return:        None
     """
